requirements.txt
asgiref==3.6.0
Django==4.1.5
python_dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
pytz==2022.7  # Corrected package name
# Core Django

# Database connectors
psycopg2-binary>=2.9  # For PostgreSQL
mysqlclient>=2.1      # For MySQL (use only if applicable)

# Django Rest Framework
djangorestframework>=3.14

# JWT Authentication
djangorestframework-simplejwt>=5.2

# CORS Headers (for API cross-origin requests)
django-cors-headers>=4.0

# Environment Variables
python-decouple>=3.8

# Image handling
Pillow>=10.0

# Storage (e.g., AWS S3)
boto3>=1.28
django-storages>=1.14

# Celery (asynchronous task queue)
celery>=5.3
redis>=5.0

# Testing
pytest>=7.4
pytest-django>=4.5

# Linting and Code Formatting
flake8>=6.0
black>=23.9
isort>=5.12

# Other utilities
requests>=2.31
drf-yasg>=1.22  # API documentation with Swagger/OpenAPI
gunicorn>=21.0  # Production WSGI server

# Optional: If using Channels (WebSockets)
channels>=4.0

# Optional: If using GraphQL
graphene-django>=3.0

# Optional: Frontend serving (if required)
whitenoise>=6.5  # For static files handling
